requirements.txt
fastapi
uvicorn
aiohttp
aiogarmin==0.1.4
python-dotenv
